rename session 2 account
