+++ packages/lib/server-side/src/index.ts
@@ -1,5 +1,5 @@
 export { Auth } from './auth';
 export * from './utils';
 export { getCachedWebProvider } from './web3Provider/getCachedWebProvider';
-export type { IAccountDatabase } from './iSessionDatabase';
+export type { IAccountDatabase } from './iAccountDatabase';
 export { Profile } from './profile';
