+++ packages/lib/server-side/src/profile.ts
@@ -3,7 +3,7 @@ import { normalizeEnsName, schema } from '@dm3-org/dm3-lib-profile';
 import { validateSchema } from '@dm3-org/dm3-lib-shared';
 import { ethers } from 'ethers';
 import express from 'express';
-import { IAccountDatabase } from './iSessionDatabase';
+import { IAccountDatabase } from './iAccountDatabase';
 
 export const Profile = (
     db: IAccountDatabase,
